update name file

+++ code.cpp
@@ -449,7 +449,7 @@ void gotoxy( int x,int y)   //hàm tọa độ
   static HANDLE h = NULL;  
   if(!h)
     h = GetStdHandle(STD_OUTPUT_HANDLE);
-  COORD c = { x, y };  
+  COORD c = { short(x), short(y) };  
   SetConsoleCursorPosition(h,c);
 };
